refactor(moyo): gather layer/magnetic data modules into submodules
Move `data/layer_*.rs` (7 files: arithmetic_crystal_class, centering,
classification, hall_symbol_database, point_group, setting, wyckoff)
into `moyo/src/data/layer/`, and `data/magnetic_*.rs` (2 files:
hall_symbol_database, space_group) into `moyo/src/data/magnetic/`.

The new `moyo/src/data/layer.rs` and `moyo/src/data/magnetic.rs` declare
the inner files as `pub(super) mod` and re-export their public items.
`data.rs` switches the matching `pub use` blocks to go through
`layer::*` and `magnetic::*`. The public API surface
(`crate::data::Layer*`, `crate::data::Magnetic*`, `UNINumber`, etc.) is
unchanged.

`hall_symbol.rs` (which crosses both groups) keeps its top-level
position; its `super::layer_*` / `super::magnetic_*` paths become
`super::layer::layer_*` / `super::magnetic::magnetic_*`.

Internal `super::*` paths in moved files become `super::super::*` for
top-level data siblings (`classification`, `hall_symbol`,
`hall_symbol_database`, `setting`); intra-submodule sibling imports
(e.g. `layer_centering` from within `layer/`) stay as `super::*`.

No behavior change; all 178 cargo tests pass.
