+++ moyo/src/data.rs
@@ -3,15 +3,8 @@ mod centering;
 mod classification;
 mod hall_symbol;
 mod hall_symbol_database;
-mod layer_arithmetic_crystal_class;
-mod layer_centering;
-mod layer_classification;
-mod layer_hall_symbol_database;
-mod layer_point_group;
-mod layer_setting;
-mod layer_wyckoff;
-mod magnetic_hall_symbol_database;
-mod magnetic_space_group;
+mod layer;
+mod magnetic;
 mod point_group;
 mod setting;
 mod wyckoff;
@@ -25,27 +18,22 @@ pub use classification::{
 };
 pub use hall_symbol::{HallSymbol, LayerHallSymbol, MagneticHallSymbol};
 pub use hall_symbol_database::{HallNumber, HallSymbolEntry, Number, hall_symbol_entry};
-pub use layer_arithmetic_crystal_class::{
-    LayerArithmeticCrystalClassEntry, LayerArithmeticNumber, iter_layer_arithmetic_crystal_entry,
-    layer_arithmetic_crystal_class_entry,
+pub use layer::{
+    LayerArithmeticCrystalClassEntry, LayerArithmeticNumber, LayerBravaisClass, LayerCentering,
+    LayerCrystalSystem, LayerHallNumber, LayerHallSymbolEntry, LayerLatticeSystem, LayerNumber,
+    LayerSetting, iter_layer_arithmetic_crystal_entry, iter_layer_hall_symbol_entry,
+    layer_arithmetic_crystal_class_entry, layer_hall_symbol_entry,
 };
-pub use layer_centering::LayerCentering;
-pub use layer_classification::{LayerBravaisClass, LayerCrystalSystem, LayerLatticeSystem};
-pub use layer_hall_symbol_database::{
-    LayerHallNumber, LayerHallSymbolEntry, LayerNumber, iter_layer_hall_symbol_entry,
-    layer_hall_symbol_entry,
-};
-pub use layer_setting::LayerSetting;
-pub use magnetic_hall_symbol_database::{MagneticHallSymbolEntry, magnetic_hall_symbol_entry};
-pub use magnetic_space_group::{
-    ConstructType, MagneticSpaceGroupType, NUM_MAGNETIC_SPACE_GROUP_TYPES, UNINumber,
-    get_magnetic_space_group_type,
+pub use magnetic::{
+    ConstructType, MagneticHallSymbolEntry, MagneticSpaceGroupType, NUM_MAGNETIC_SPACE_GROUP_TYPES,
+    UNINumber, get_magnetic_space_group_type, magnetic_hall_symbol_entry,
 };
 pub use setting::Setting;
 
 pub(super) use arithmetic_crystal_class::iter_arithmetic_crystal_entry;
-pub(super) use layer_point_group::LayerPointGroupRepresentative;
-pub(super) use layer_wyckoff::{LayerWyckoffPosition, iter_layer_wyckoff_positions};
-pub(super) use magnetic_space_group::uni_number_range;
+pub(super) use layer::{
+    LayerPointGroupRepresentative, LayerWyckoffPosition, iter_layer_wyckoff_positions,
+};
+pub(super) use magnetic::uni_number_range;
 pub(super) use point_group::PointGroupRepresentative;
 pub(super) use wyckoff::{WyckoffPosition, WyckoffPositionSpace, iter_wyckoff_positions};
