+++ moyo/src/data/magnetic.rs
@@ -1,0 +1,10 @@
+pub(super) mod magnetic_hall_symbol_database;
+pub(super) mod magnetic_space_group;
+
+pub use magnetic_hall_symbol_database::{MagneticHallSymbolEntry, magnetic_hall_symbol_entry};
+pub use magnetic_space_group::{
+    ConstructType, MagneticSpaceGroupType, NUM_MAGNETIC_SPACE_GROUP_TYPES, UNINumber,
+    get_magnetic_space_group_type,
+};
+
+pub use magnetic_space_group::uni_number_range;
